requirements.txt
Django==1.11.2
Pillow==4.1.1
django-simple-captcha==0.5.5
djangorestframework==3.6.3
olefile==0.44
pytz==2017.2
six==1.10.0
